use file icon
